requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
ffmpeg-python==0.2.0
frozenlist==1.8.0
future==1.0.0
idna==3.11
multidict==6.7.0
nodejs-wheel-binaries==24.12.0
propcache==0.4.1
pyfiglet==1.0.4
pytubefix==10.3.6
termcolor==3.2.0
yarl==1.22.0
